requirements.txt
pyyaml
PyInquirer